Resolver source codes reformatted.

+++ resolver/chptr.c
@@ -12,15 +12,15 @@
  * Copyright (c) 2015 Parham Alvani.
 */
 #include <stdint.h>
+#include <stddef.h>
 
 #include "chptr.h"
-#include "common.h"
 
 uint32_t chptr_checksum(const struct chptrhdr *chptr)
 {
 	uint64_t sum = 0;
 	uint32_t checksum = 0;
-	int len = sizeof(struct chptrhdr);
+	size_t len = sizeof(struct chptrhdr);
 	const uint32_t *chptr_header = (const uint32_t *) chptr;
 
 	while (len > 1) {
@@ -28,7 +28,7 @@ uint32_t chptr_checksum(const struct chptrhdr *chptr)
 		len -= sizeof(uint32_t);
 	}
 	if (len)
-		sum += *((uint16_t *)(chptr_header));
+		sum += *((uint16_t *) (chptr_header));
 
 	/* handle sum carry !! :-) this is good !! */
 	sum = (sum >> 32) + (sum & 0xFFFFFFFF);
